Move discovery output helpers into partial
