+++ src/CodeIndex/Cli/QueryCommandRunner.Discovery.cs
@@ -1,6 +1,9 @@
 using System.Globalization;
+using System.Text;
 using System.Text.Json;
+using System.Text.Json.Nodes;
 using CodeIndex.Database;
+using CodeIndex.Models;
 
 namespace CodeIndex.Cli;
 
@@ -404,5 +407,292 @@ public static partial class QueryCommandRunner
             }
             return CommandExitCodes.Success;
         });
+    }
+
+    /// <summary>
+    /// Build the OR-joined name list for `symbols`: first positional + extra positionals + --name values.
+    /// Pipe characters are treated as literal name characters so operator symbols like `operator |` remain searchable.
+    /// Multi-name queries must use repeated positional args or `--name` flags.
+    /// `symbols` コマンド用の名前リストを組み立て（最初の positional + 追加 positional + --name）。
+    /// `|` は名前文字として扱うので `operator |` などの演算子シンボルも検索可能。複数名指定は繰り返し positional か `--name` で行う。
+    /// </summary>
+    internal static (List<string>? Queries, bool HadExplicitInput) BuildSymbolQueryList(QueryCommandOptions options)
+    {
+        var raw = new List<string>();
+        if (options.Query != null)
+            raw.Add(options.Query);
+        raw.AddRange(options.ExtraNames);
+        var hadExplicitInput = raw.Count > 0;
+        if (!hadExplicitInput)
+            return (null, false);
+        var deduped = raw.Where(s => !string.IsNullOrEmpty(s)).Distinct().ToList();
+        if (deduped.Any(IsBareVerbatimQueryToken))
+            return (null, hadExplicitInput);
+        return (deduped.Count == 0 ? null : deduped, hadExplicitInput);
+    }
+
+    private sealed record FileListScopeFilters(
+        IReadOnlyList<string> PathPatterns,
+        IReadOnlyList<string> ExcludePaths,
+        bool ExcludeTests);
+
+    private static FileListScopeFilters BuildFilesScopeFilters(QueryCommandOptions options)
+    {
+        if (!options.ExcludeTests || options.PathPatterns.Count > 0)
+        {
+            return new(
+                options.PathPatterns,
+                options.ExcludePaths,
+                options.ExcludeTests);
+        }
+
+        var pathPatterns = new List<string>(options.PathPatterns);
+        AddDistinct(pathPatterns, SearchAuditRecipes.DefaultSourcePathPatterns);
+        var excludePaths = new List<string>(options.ExcludePaths);
+        AddDistinct(excludePaths, SearchAuditRecipes.DefaultSourceExcludePaths);
+        return new(pathPatterns, excludePaths, ExcludeTests: true);
+    }
+
+    private static void AddFileCountBytesJsonFields(JsonObject payload, QueryCountResult counts)
+    {
+        payload["total_bytes"] = counts.TotalBytes ?? 0;
+        payload["average_bytes"] = counts.AverageBytes ?? 0;
+        payload["max_bytes"] = counts.MaxBytes ?? 0;
+        payload["max_bytes_path"] = counts.MaxBytesPath;
+        payload["bytes_authoritative"] = counts.BytesAuthoritative ?? true;
+    }
+
+    private static string FormatFileCountBytesSummary(QueryCountResult counts)
+    {
+        var totalBytes = counts.TotalBytes ?? 0;
+        var averageBytes = counts.AverageBytes ?? 0;
+        var maxBytes = counts.MaxBytes ?? 0;
+        var maxPath = string.IsNullOrEmpty(counts.MaxBytesPath)
+            ? "none"
+            : counts.MaxBytesPath;
+        var authority = (counts.BytesAuthoritative ?? true) ? "true" : "false";
+        return string.Create(
+            CultureInfo.InvariantCulture,
+            $"{counts.Count} files, {totalBytes} bytes total, average {averageBytes:0.##} bytes, max {maxBytes} bytes ({maxPath}), bytes_authoritative: {authority}");
+    }
+
+    private static bool ShouldWriteBoundedDiscoveryJsonPayload(QueryCommandOptions options)
+        => options.OutputFormat == OutputFormatCompact || options.SummaryOnly || options.MaxJsonBytes.HasValue;
+
+    private static bool TryWriteDiscoveryOutputControlUsageError(string commandName, QueryCommandOptions options)
+    {
+        if (!options.SummaryOnly && !options.MaxJsonBytes.HasValue)
+            return false;
+        if (options.Json && options.OutputFormat is OutputFormatJson or OutputFormatCompact or OutputFormatCount)
+            return false;
+
+        var control = options.SummaryOnly ? "--summary-only" : "--max-json-bytes";
+        WriteUsageError(
+            $"{control} is only supported with {commandName} JSON, compact, or count output.",
+            GetUsageLineOrThrow(commandName),
+            $"Use `cdidx {commandName} --json {control}` or `cdidx {commandName} --format compact {control}`.");
+        return true;
+    }
+
+    private static int WriteBoundedDiscoveryJsonPayload<T>(
+        DbReader reader,
+        QueryCommandOptions options,
+        JsonSerializerOptions jsonOptions,
+        string commandName,
+        string resultsKey,
+        IReadOnlyList<T> results,
+        int totalCount,
+        int fileCount,
+        Func<T, JsonNode?> rowFactory,
+        ExactQuerySignal? exactSignal = null)
+    {
+        var jsonNodeOptions = EnsureJsonNodeSerializerOptions(jsonOptions);
+        var requestedRows = options.SummaryOnly ? 0 : results.Count;
+        var json = BuildBoundedDiscoveryJson(requestedRows);
+        if (!options.MaxJsonBytes.HasValue)
+        {
+            Console.WriteLine(json);
+            return CommandExitCodes.Success;
+        }
+
+        if (JsonFitsByteLimit(json, options.MaxJsonBytes.Value))
+        {
+            Console.WriteLine(json);
+            return CommandExitCodes.Success;
+        }
+
+        string? bestJson = null;
+        var low = 0;
+        var high = requestedRows;
+        while (low <= high)
+        {
+            var mid = low + ((high - low) / 2);
+            var candidate = BuildBoundedDiscoveryJson(mid);
+            if (JsonFitsByteLimit(candidate, options.MaxJsonBytes.Value))
+            {
+                bestJson = candidate;
+                low = mid + 1;
+            }
+            else
+            {
+                high = mid - 1;
+            }
+        }
+
+        if (bestJson != null)
+        {
+            Console.WriteLine(bestJson);
+            return CommandExitCodes.Success;
+        }
+
+        return WriteJsonObjectWithOptionalByteLimit(
+            BuildBoundedDiscoveryJson(0),
+            options,
+            $"{commandName} compact",
+            "Use --summary-only, reduce --limit, or increase --max-json-bytes.",
+            commandName);
+
+        string BuildBoundedDiscoveryJson(int emittedRows)
+        {
+            var payload = BuildBoundedDiscoveryPayload(
+                reader,
+                options,
+                jsonOptions,
+                resultsKey,
+                results,
+                totalCount,
+                fileCount,
+                Math.Clamp(emittedRows, 0, results.Count),
+                rowFactory,
+                exactSignal);
+            return payload.ToJsonString(jsonNodeOptions);
+        }
+    }
+
+    private static JsonObject BuildBoundedDiscoveryPayload<T>(
+        DbReader reader,
+        QueryCommandOptions options,
+        JsonSerializerOptions jsonOptions,
+        string resultsKey,
+        IReadOnlyList<T> results,
+        int totalCount,
+        int fileCount,
+        int emittedRows,
+        Func<T, JsonNode?> rowFactory,
+        ExactQuerySignal? exactSignal)
+    {
+        var omittedCount = Math.Max(0, totalCount - emittedRows);
+        var rowLimitReached = totalCount > results.Count;
+        var byteLimitReached = !options.SummaryOnly && options.MaxJsonBytes.HasValue && emittedRows < results.Count;
+        var payload = new JsonObject
+        {
+            ["api_version"] = JsonOutputContract.ApiVersion,
+            ["count"] = totalCount,
+            ["file_count"] = fileCount,
+            ["emitted_count"] = emittedRows,
+            ["omitted_count"] = omittedCount,
+            ["truncated"] = omittedCount > 0,
+        };
+
+        if (options.OutputFormat == OutputFormatCompact)
+            payload["format"] = OutputFormatCompact;
+        if (options.SummaryOnly)
+            payload["summary_only"] = true;
+        if (options.MaxJsonBytes.HasValue)
+            payload["max_json_bytes"] = options.MaxJsonBytes.Value;
+        if (rowLimitReached)
+            payload["row_limit_reached"] = true;
+        if (byteLimitReached)
+            payload["byte_limit_reached"] = true;
+
+        var omittedBy = new JsonArray();
+        if (options.SummaryOnly && totalCount > 0)
+            omittedBy.Add("summary_only");
+        if (rowLimitReached)
+            omittedBy.Add("limit");
+        if (byteLimitReached)
+            omittedBy.Add("max_json_bytes");
+        if (omittedBy.Count > 0)
+            payload["omitted_by"] = omittedBy;
+
+        if (!options.SummaryOnly)
+            payload[resultsKey] = BuildDiscoveryRows(results, emittedRows, rowFactory);
+        if (exactSignal.HasValue)
+            AddExactJsonFields(payload, exactSignal.Value);
+        payload["query_context"] = BuildQueryContextJson(options, jsonOptions);
+        AddFreshnessHint(payload, reader);
+        return payload;
+    }
+
+    private static JsonArray BuildDiscoveryRows<T>(IReadOnlyList<T> results, int emittedRows, Func<T, JsonNode?> rowFactory)
+    {
+        var rows = new JsonArray();
+        for (var i = 0; i < emittedRows && i < results.Count; i++)
+            rows.Add(rowFactory(results[i]));
+        return rows;
+    }
+
+    private static bool JsonFitsByteLimit(string json, int maxJsonBytes)
+        => Encoding.UTF8.GetByteCount(json) + Environment.NewLine.Length <= maxJsonBytes;
+
+    private static JsonNode? ToFileDiscoveryJsonNode(FileResult result, JsonSerializerOptions jsonOptions, bool compact)
+    {
+        if (!compact)
+            return JsonSerializer.SerializeToNode(result, CliJsonSerializerContextFactory.Create(jsonOptions).FileResult);
+
+        var row = new JsonObject
+        {
+            ["path"] = result.Path,
+            ["lines"] = result.Lines,
+            ["size"] = result.Size,
+            ["symbol_count"] = result.SymbolCount,
+            ["reference_count"] = result.ReferenceCount,
+        };
+        if (result.Lang != null)
+            row["lang"] = result.Lang;
+        return row;
+    }
+
+    private static JsonNode? ToSymbolDiscoveryJsonNode(SymbolResult result, JsonSerializerOptions jsonOptions, bool compact)
+    {
+        if (!compact)
+            return JsonSerializer.SerializeToNode(result, CliJsonSerializerContextFactory.Create(jsonOptions).SymbolResult);
+
+        var row = new JsonObject
+        {
+            ["path"] = result.Path,
+            ["line"] = result.Line,
+            ["start_line"] = result.StartLine,
+            ["end_line"] = result.EndLine,
+            ["kind"] = result.Kind,
+            ["name"] = result.Name,
+        };
+        if (result.Lang != null)
+            row["lang"] = result.Lang;
+        if (result.ContainerName != null)
+            row["container_name"] = result.ContainerName;
+        if (result.Visibility != null)
+            row["visibility"] = result.Visibility;
+        if (result.SortMode != null)
+            row["sort_mode"] = result.SortMode;
+        if (result.ReferenceCount.HasValue)
+            row["reference_count"] = result.ReferenceCount.Value;
+        if (result.HotspotScore.HasValue)
+            row["hotspot_score"] = result.HotspotScore.Value;
+        if (result.RankingReferenceScore.HasValue)
+            row["ranking_reference_score"] = result.RankingReferenceScore.Value;
+        if (result.RankingHotspotScore.HasValue)
+            row["ranking_hotspot_score"] = result.RankingHotspotScore.Value;
+        if (result.GenericNamePenalty.HasValue)
+            row["generic_name_penalty"] = result.GenericNamePenalty.Value;
+        if (result.StructuralRankPenalty.HasValue)
+            row["structural_rank_penalty"] = result.StructuralRankPenalty.Value;
+        if (result.DefinitionSites.HasValue)
+            row["definition_sites"] = result.DefinitionSites.Value;
+        if (result.SizeLines.HasValue)
+            row["size_lines"] = result.SizeLines.Value;
+        if (result.ComplexityScore.HasValue)
+            row["complexity_score"] = result.ComplexityScore.Value;
+        return row;
     }
 }
